estilos header post

+++ src/model/controller-post.js
@@ -1,9 +1,8 @@
-/* eslint-disable import/no-cycle */
-// import { dataBase } from '../main.js';
+
 
 export const datePost = () => {
   const opt1 = {
-    month: 'long', day: 'numeric', year: 'numeric',
+    month: 'short', day: 'numeric', year: 'numeric',
   };
   const opt2 = {
     hour12: 'true', hour: 'numeric', minute: 'numeric',
